s1build 1/5: widen the DFA_SELECT deny plumbing to uint64_t (litscan_s1.md R4/R3-7)
Every site the grep recipe `unsigned .*= .*opt->flags` finds, plus the
typed fields and parameters it feeds: DfaCand.deny, dfa_select's flags,
dfa_form_derive's local, compile.c's pfc_flags, PcrecAxisCand.deny, and
axes_dump.c's axis_macro_name/axis_cli_flag/bit_of/deny_cols/PredAxis/
emit_pred_row. No bit above 31 exists yet; this lands BEFORE the bit-32
#define (R3-7's ordering rule). Zero artifacts, --list-axes byte-identical.

Adds docs/dev/optloop/s1/s1_identity.py, the per-commit identity gate.

+++ src/core/internal.h
@@ -6217,7 +6217,7 @@ int pcrec_req_byte(Ctx *cx, const Ast *root, bool run_ok, ReqRun *run,
  * bits (docs/dev/learnings.md §3) — so a candidate added to one of those six
  * lists appears in the dump with no edit to the walker. `cap` bounds `out`;
  * returns the number written (never more than `cap`). */
-typedef struct { const char *name; unsigned deny; } PcrecAxisCand;
+typedef struct { const char *name; uint64_t deny; } PcrecAxisCand;
 
 size_t pcrec_dfa_axis_table_cands(PcrecAxisCand *out, size_t cap);      /* axis A */
